pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conversion-engine"
version = "0.1.0"
description = "Pluggable conversion engine scaffold for multi-client outbound"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.28.1",
  "requests>=2.32.3",
  "playwright>=1.54.0",
  "langfuse>=2.60.0",
  "hubspot-api-client>=11.2.0",
]

[tool.setuptools]
packages = [
  "agent",
  "agent.policies",
  "channels",
  "channels.email",
  "channels.sms",
  "channels.voice",
  "enrichment",
  "enrichment.tenacious",
  "enrichment.acme",
  "crm",
  "calendars",
  "probes",
  "evidence",
]
